backend/requirements.txt
# =============================================================================
# AI Research Agent — Python Dependencies
# =============================================================================

# --- Core Framework ---
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
pydantic==2.11.3

# --- LLM & Agent Orchestration ---
langchain==0.3.25
langchain-google-genai==2.1.4
langgraph==0.4.7
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langchain-chroma==0.2.4

# --- Vector Database ---
chromadb>=1.0.9

# --- Web Search ---
tavily-python==0.5.0

# --- Document Processing ---
pypdf==5.4.0
python-multipart==0.0.20

# --- Utilities ---
aiofiles==24.1.0
websockets==15.0.1
httpx==0.28.1

# --- Development & Quality ---
ruff==0.11.8
pytest==8.3.5
pytest-asyncio==0.26.0
